TableView - not displaying Firebase
Strange issue where the Firebase data isn't being displayed. There's a function with dummy data that works completely okay and populates the table as it should but when calling anything from Firebase: nothing appears at all.

There's a function to ensure that the app writes to Firebase okay and it shows on the console but said entries are never shown.