TableView issue
Can easily write to the Firebase db but have issues showing the issues on the TableView from the DB.

Dummy data made within the app can be displayed with no issues and able to write entries to the DB too but an issue with displaying. Will pour over again and see where the issue lies.